+++ activerecord/lib/active_record/connection_adapters/abstract/query_cache.rb
@@ -21,7 +21,9 @@ module ActiveRecord
           method_names.each do |method_name|
             base.class_eval <<-end_code, __FILE__, __LINE__ + 1
               def #{method_name}(...)
-                ActiveRecord::Base.clear_query_caches_for_current_thread
+                if query_cache_enabled || pool.query_cache_enabled
+                  ActiveRecord::Base.clear_query_caches_for_current_thread
+                end
                 super
               end
             end_code